Extract CssTemplate-related views to their own file from core.py
